removed deprecated option ngrt4nd and added basis constants for Kubernetes support

+++ core/src/utilsCore.hpp
@@ -149,7 +149,7 @@ namespace ngrt4n
 
   QString sourceData2Json(const SourceT& src);
 
-  qint32 convertToSourceType(const QString& str);
+  qint8 convertToSourceType(const QString& str);
 
   void setCheckOnError(int status, const QString& msg, CheckT& invalidCheck);
